refactor email portal settings
